[SAW-XXX] Make cli timeout float

+++ src/test_farm/cli.py
@@ -22,7 +22,7 @@ def run(
         ..., dir_okay=False, readable=True, resolve_path=True
     ),
     controller_bind_address: str = typer.Option(..., "--controller-bind-address"),
-    receipt_timeout_seconds: int = typer.Option(0, "--receipt-timeout-seconds", min=0),
+    receipt_timeout_seconds: float = typer.Option(0.0, "--receipt-timeout-seconds", min=0.0),
     results_dir: Path = typer.Option(
         Path("results"), "--results-dir", file_okay=False, resolve_path=True
     ),
